requirements.txt
torch
torchvision
timm
albumentations
opencv-python
albumentations[imgaug]
pillow
scikit-learn
streamlit
tqdm
